DRF_test/requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
